[bluelink] Allow for double heading values

+++ bundles/org.openhab.binding.bluelink/src/main/java/org/openhab/binding/bluelink/internal/dto/eu/ccs2/Ccs2VehicleStatusResponse.java
@@ -251,7 +251,7 @@ public record Ccs2VehicleStatusResponse(String resCode, @SerializedName("Service
 
     // Location & Coordinates
     public record VehicleLocation(@SerializedName("Data") String time, @SerializedName("GeoCoord") GeoCoordinates coord,
-            @SerializedName("Heading") int head, @SerializedName("Speed") ValueUnit speed) {
+            @SerializedName("Heading") double head, @SerializedName("Speed") ValueUnit speed) {
     }
 
     public record GeoCoordinates(@SerializedName("Latitude") double lat, @SerializedName("Longitude") double lon,
